requirements.txt
Pillow==10.3.0
google-generativeai==0.7.2
line-bot-sdk==3.11.0
requests==2.31.0
